event-generator/requirements.txt
kafka-python==2.0.2
aiochclient[aiohttp]==2.6.0
pydantic==2.5.3
pydantic-settings==2.1.0
sqlalchemy-utils==0.41.2
asyncpg==0.29.0